Move persistence into separate file
